requirements.txt
python-pptx
Pillow 
pandas
ipykernel
requests
gdown
PyMuPDF
pyyaml